+++ src/main/java/duke/DeleteCommand.java
@@ -1,7 +1,9 @@
-public class DoneCommand extends Command {
+package duke;
+
+public class DeleteCommand extends Command {
     private int index;
 
-    public DoneCommand(int index) {
+    public DeleteCommand(int index) {
         super();
         this.index = index;
     }
@@ -11,8 +13,10 @@ public class DoneCommand extends Command {
         if (index > tasks.getSize() || index <= 0) {
             throw new DukeException("OOPS!!! I'm sorry, but you've entered an invalid index.");
         }
-        Task t = tasks.markAsDone(index);
+        Task t = tasks.delete(index);
         storage.save(tasks);
-        ui.showResponse("Nice! I've marked this task as done: \n\t\t " + t.toString());
+        ui.showResponse("Noted. I've removed this task: \n\t\t "
+                + t
+                + "\n\t Now you have " + tasks.getSize() + " tasks in the list.");
     }
 }
