+++ src/main/java/duke/DoneCommand.java
@@ -1,7 +1,9 @@
-public class DeleteCommand extends Command {
+package duke;
+
+public class DoneCommand extends Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public DoneCommand(int index) {
         super();
         this.index = index;
     }
@@ -11,10 +13,8 @@ public class DeleteCommand extends Command {
         if (index > tasks.getSize() || index <= 0) {
             throw new DukeException("OOPS!!! I'm sorry, but you've entered an invalid index.");
         }
-        Task t = tasks.delete(index);
+        Task t = tasks.markAsDone(index);
         storage.save(tasks);
-        ui.showResponse("Noted. I've removed this task: \n\t\t "
-                + t
-                + "\n\t Now you have " + tasks.getSize() + " tasks in the list.");
+        ui.showResponse("Nice! I've marked this task as done: \n\t\t " + t.toString());
     }
 }
