+++ src/main/java/duke/Duke.java
@@ -1,11 +1,4 @@
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Scanner;
+package duke;
 
 public class Duke {
     private Storage storage;
